requirements.txt
mysql-connector-python
requests
PyYAML
click 